+++ src/plugin/Dir.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "../Plugin.h"
 
-extern Plugin Ls_Plugin;
+extern Plugin Dir_Plugin;