+++ test/unsupported_input.t
@@ -4,6 +4,6 @@ Setup:
 
 Complain about non-supported input:
 
-  $ show -d file -d ls -d pager -d pipe doesnt_exist
+  $ show -d file -d dir -d pager -d pipe doesnt_exist
   Unsupported input: doesnt_exist
   [1]